+++ test/classes/StorageEngineTest.php
@@ -1,4 +1,5 @@
 <?php
+use PMA\libraries\StorageEngine;
 /**
  * Tests for StorageEngine.php
  *
@@ -100,16 +101,43 @@ class StorageEngineTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-     * Test for getEngine
+     * Test for StorageEngine::getEngine
+     *
+     * @param string $expectedClass Class that should be selected
+     * @param string $engineName    Engine name
      *
      * @return void
+     *
+     * @dataProvider providerGetEngine
      */
-    public function testGetEngine()
+    public function testGetEngine($expectedClass, $engineName)
     {
-
         $this->assertInstanceOf(
-            'PMA\libraries\StorageEngine',
-            $this->object->getEngine('dummy')
+            $expectedClass, StorageEngine::getEngine($engineName)
+        );
+    }
+
+    /**
+     * Provider for testGetEngine
+     *
+     * @return array
+     */
+    public function providerGetEngine()
+    {
+        return array(
+            array('PMA\libraries\StorageEngine', 'unknown engine'),
+            array('PMA\libraries\engines\Bdb', 'Bdb'),
+            array('PMA\libraries\engines\Berkeleydb', 'Berkeleydb'),
+            array('PMA\libraries\engines\Binlog', 'Binlog'),
+            array('PMA\libraries\engines\Innobase', 'Innobase'),
+            array('PMA\libraries\engines\Innodb', 'Innodb'),
+            array('PMA\libraries\engines\Memory', 'Memory'),
+            array('PMA\libraries\engines\Merge', 'Merge'),
+            array('PMA\libraries\engines\Mrg_Myisam', 'Mrg_Myisam'),
+            array('PMA\libraries\engines\Myisam', 'Myisam'),
+            array('PMA\libraries\engines\Ndbcluster', 'Ndbcluster'),
+            array('PMA\libraries\engines\Pbxt', 'Pbxt'),
+            array('PMA\libraries\engines\Performance_Schema', 'Performance_Schema'),
         );
     }
